requirements.txt
# Core
Django==5.2.10
sc4py==0.1.4

# Models
django-better-choices==1.18
django-polymorphic==4.10.2

# Rule engine
jsonschema==4.26.0
rule-engine==4.5.3
-e ../django_rule_engine

# REST
httpie==3.2.4
requests==2.32.5
# django-ninja==1.5.3

# Cache and Session
django-valkey==0.4.0
valkey==6.1.1

# # Security
# cryptography==46.0.3
# pyjwt==2.10.1

# Audit
django-simple-history==3.11.0
# django-safedelete==1.4.1

# UI
whitenoise==6.11.0
django-import-export==4.4.0
django-json-widget==2.1.1
django-admin-autocomplete-filter==0.7.1
django-richtextfield==1.6.2

# Database
psycopg[binary,pool]==3.3.2

# Production
sentry-sdk[django,rq]==2.49.0
gunicorn==23.0.0
